requirements.txt
twelvelabs
streamlit
fastapi
uvicorn
faiss-cpu
numpy
python-dotenv
ipykernel